refactor: split pkg/runtime/docker/docker.go - client.go + inspect.go (issue 3, part A)
