+++ pkg/runtime/docker/docker.go
@@ -3,7 +3,6 @@ package docker
 import (
 	"bytes"
 	"context"
-	"crypto/tls"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -14,59 +13,15 @@ import (
 
 	ctr "github.com/alexei-led/pumba/pkg/container"
 	ctypes "github.com/docker/docker/api/types/container"
-	"github.com/docker/docker/api/types/filters"
 	imagetypes "github.com/docker/docker/api/types/image"
-	dockerapi "github.com/docker/docker/client"
 	"github.com/docker/go-connections/nat"
 	log "github.com/sirupsen/logrus"
 )
-
-// NewClient returns a new Client instance which can be used to interact with the Docker API.
-func NewClient(dockerHost string, tlsConfig *tls.Config) (ctr.Client, error) {
-	apiClient, err := NewAPIClient(dockerHost, tlsConfig)
-	if err != nil {
-		return nil, err
-	}
-	return NewFromAPI(apiClient)
-}
-
-// NewAPIClient returns a bare Docker SDK client. Exposed so alternate runtimes
-// (e.g. Podman via the Docker-compat socket) can reuse the HTTP/TLS setup.
-func NewAPIClient(dockerHost string, tlsConfig *tls.Config) (*dockerapi.Client, error) {
-	httpClient, err := HTTPClient(dockerHost, tlsConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	apiClient, err := dockerapi.NewClientWithOpts(dockerapi.WithHost(dockerHost), dockerapi.WithHTTPClient(httpClient), dockerapi.WithAPIVersionNegotiation())
-	if err != nil {
-		return nil, fmt.Errorf("failed to create docker client: %w", err)
-	}
-	return apiClient, nil
-}
-
-// NewFromAPI wraps an existing Docker SDK client as a ctr.Client. Exposed so
-// alternate runtimes can reuse the Docker implementation via embedding.
-func NewFromAPI(api *dockerapi.Client) (ctr.Client, error) {
-	if api == nil {
-		return nil, errors.New("docker: api client must not be nil")
-	}
-	return dockerClient{containerAPI: api, imageAPI: api, systemAPI: api}, nil
-}
 
 const (
 	defaultStopSignal = "SIGTERM"
 	defaultKillSignal = "SIGKILL"
 )
-
-type dockerClient struct {
-	containerAPI dockerapi.ContainerAPIClient
-	imageAPI     dockerapi.ImageAPIClient
-	systemAPI    dockerapi.SystemAPIClient
-}
-
-// Close is a no-op for the Docker client; the underlying HTTP connections are managed by the SDK.
-func (client dockerClient) Close() error { return nil }
 
 type imagePullResponse struct {
 	Status         string `json:"status"`
@@ -78,76 +33,8 @@ type imagePullResponse struct {
 	} `json:"progressDetail"`
 }
 
-// dockerInspectToContainer converts Docker inspect responses into a runtime-agnostic Container.
-func dockerInspectToContainer(info ctypes.InspectResponse, img *imagetypes.InspectResponse) *ctr.Container {
-	c := &ctr.Container{
-		ImageID:  img.ID,
-		Labels:   make(map[string]string),
-		Networks: make(map[string]ctr.NetworkLink),
-	}
-	if info.ContainerJSONBase != nil {
-		c.ContainerID = info.ID
-		c.ContainerName = info.Name
-		c.Image = info.Image
-		if info.State != nil {
-			if info.State.Running {
-				c.State = ctr.StateRunning
-			} else {
-				c.State = ctr.StateExited
-			}
-		}
-	}
-	if info.Config != nil && info.Config.Labels != nil {
-		c.Labels = info.Config.Labels
-	}
-	if info.NetworkSettings != nil {
-		for name, ep := range info.NetworkSettings.Networks {
-			c.Networks[name] = ctr.NetworkLink{Links: ep.Links}
-		}
-	}
-	return c
-}
-
 const cgroupDriverSystemd = "systemd"
 const cgroupDriverCgroupfs = "cgroupfs"
-
-// ListContainers returns a list of containers that match the given filter
-func (client dockerClient) ListContainers(ctx context.Context, fn ctr.FilterFunc, opts ctr.ListOpts) ([]*ctr.Container, error) {
-	filterArgs := filters.NewArgs()
-	for _, label := range opts.Labels {
-		filterArgs.Add("label", label)
-	}
-	return client.listContainers(ctx, fn, ctypes.ListOptions{All: opts.All, Filters: filterArgs})
-}
-
-func (client dockerClient) listContainers(ctx context.Context, fn ctr.FilterFunc, opts ctypes.ListOptions) ([]*ctr.Container, error) {
-	log.Debug("listing containers")
-	containers, err := client.containerAPI.ContainerList(ctx, opts)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list containers: %w", err)
-	}
-	var cs []*ctr.Container
-	for _, container := range containers {
-		containerInfo, err := client.containerAPI.ContainerInspect(ctx, container.ID)
-		if err != nil {
-			return nil, fmt.Errorf("failed to inspect container: %w", err)
-		}
-		log.WithFields(log.Fields{
-			"name": containerInfo.Name,
-			"id":   containerInfo.ID,
-		}).Debug("found container")
-
-		imgInfo, err := client.imageAPI.ImageInspect(ctx, containerInfo.Image)
-		if err != nil {
-			log.WithError(err).WithField("image", containerInfo.Image).Warn("failed to inspect container image, skipping image metadata")
-		}
-		c := dockerInspectToContainer(containerInfo, &imgInfo)
-		if fn(c) {
-			cs = append(cs, c)
-		}
-	}
-	return cs, nil
-}
 
 // KillContainer kills a container with the given signal
 func (client dockerClient) KillContainer(ctx context.Context, c *ctr.Container, signal string, dryrun bool) error {
