+++ pkg/runtime/docker/docker_test.go
@@ -51,104 +51,6 @@ func mockNoContainers(_ *ctr.Container) bool {
 	return false
 }
 
-func TestListContainers_Success(t *testing.T) {
-	containerDetails := DetailsResponse(AsMap("ID", "foo", "Name", "/bar", "Image", "abc123"))
-	allContainersResponse := Containers(Response(AsMap(
-		"ID", "foo",
-		"Names", []string{"bar"})),
-	)
-	imageDetails := ImageDetailsResponse(AsMap("ID", "abc123"))
-
-	api := NewMockEngine()
-	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(allContainersResponse, nil)
-	api.EXPECT().ContainerInspect(mock.Anything, "foo").Return(containerDetails, nil)
-	api.EXPECT().ImageInspect(mock.Anything, "abc123").Return(imageDetails, nil)
-
-	client := dockerClient{containerAPI: api, imageAPI: api}
-
-	// Verify that dockerClient implements container.Client interface
-	var _ ctr.Client = (*dockerClient)(nil)
-
-	containers, err := client.ListContainers(context.TODO(), mockAllContainers, ctr.ListOpts{All: true})
-
-	assert.NoError(t, err)
-	assert.Len(t, containers, 1)
-	assert.Equal(t, "foo", containers[0].ID())
-	assert.Equal(t, "/bar", containers[0].Name())
-	assert.Equal(t, "abc123", containers[0].ImageID)
-	api.AssertExpectations(t)
-}
-
-func TestListContainers_Filter(t *testing.T) {
-	containerDetails := DetailsResponse(AsMap("ID", "foo", "Name", "/bar", "Image", "abc123"))
-	allContainersResponse := Containers(Response(AsMap(
-		"ID", "foo",
-		"Names", []string{"bar"})),
-	)
-	imageDetails := ImageDetailsResponse(AsMap("ID", "abc123"))
-
-	api := NewMockEngine()
-	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(allContainersResponse, nil)
-	api.EXPECT().ContainerInspect(mock.Anything, "foo").Return(containerDetails, nil)
-	api.EXPECT().ImageInspect(mock.Anything, "abc123").Return(imageDetails, nil)
-
-	client := dockerClient{containerAPI: api, imageAPI: api}
-	containers, err := client.ListContainers(context.TODO(), mockNoContainers, ctr.ListOpts{})
-
-	assert.NoError(t, err)
-	assert.Len(t, containers, 0)
-	api.AssertExpectations(t)
-}
-
-func TestListContainers_ListError(t *testing.T) {
-	api := NewMockEngine()
-	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(Containers(), errors.New("oops"))
-
-	client := dockerClient{containerAPI: api, imageAPI: api}
-	_, err := client.ListContainers(context.TODO(), mockAllContainers, ctr.ListOpts{All: true})
-
-	assert.Error(t, err)
-	assert.EqualError(t, err, "failed to list containers: oops")
-	api.AssertExpectations(t)
-}
-
-func TestListContainers_InspectContainerError(t *testing.T) {
-	api := NewMockEngine()
-	allContainersResponse := Containers(Response(AsMap(
-		"ID", "foo",
-		"Names", []string{"bar"})),
-	)
-	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(allContainersResponse, nil)
-	api.EXPECT().ContainerInspect(mock.Anything, "foo").Return(DetailsResponse(AsMap()), errors.New("uh-oh"))
-
-	client := dockerClient{containerAPI: api, imageAPI: api}
-	_, err := client.ListContainers(context.TODO(), mockAllContainers, ctr.ListOpts{All: true})
-
-	assert.Error(t, err)
-	assert.EqualError(t, err, "failed to inspect container: uh-oh")
-	api.AssertExpectations(t)
-}
-
-func TestListContainers_InspectImageError(t *testing.T) {
-	allContainersResponse := Containers(Response(AsMap(
-		"ID", "foo",
-		"Names", []string{"bar"})),
-	)
-	resp := DetailsResponse(AsMap("Image", "abc123"))
-	imageDetailsResponse := ImageDetailsResponse(AsMap())
-	api := NewMockEngine()
-	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(allContainersResponse, nil)
-	api.EXPECT().ContainerInspect(mock.Anything, "foo").Return(resp, nil)
-	api.EXPECT().ImageInspect(mock.Anything, "abc123").Return(imageDetailsResponse, errors.New("whoops"))
-
-	client := dockerClient{containerAPI: api, imageAPI: api}
-	containers, err := client.ListContainers(context.TODO(), mockAllContainers, ctr.ListOpts{All: true})
-
-	assert.NoError(t, err)
-	assert.Len(t, containers, 1)
-	api.AssertExpectations(t)
-}
-
 func TestStopContainer_DefaultSuccess(t *testing.T) {
 	c := NewTestContainer(AsMap("ID", "abc123", "Name", "foo", "Image", "abc123"))
 	notRunningContainer := DetailsResponse(AsMap("Running", false))
@@ -2430,55 +2332,4 @@ func TestIPTablesForSimpleCases(t *testing.T) {
 		err := client.ipTablesExecCommand(ctx, "container-id", []string{"-L"})
 		assert.NoError(t, err)
 	})
-}
-
-func TestNewAPIClient(t *testing.T) {
-	tests := []struct {
-		name    string
-		host    string
-		wantErr bool
-	}{
-		{name: "unix socket", host: "unix:///var/run/docker.sock", wantErr: false},
-		{name: "tcp host", host: "tcp://localhost:2375", wantErr: false},
-		{name: "invalid scheme url", host: "://bad", wantErr: true},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			api, err := NewAPIClient(tt.host, nil)
-			if tt.wantErr {
-				assert.Error(t, err)
-				assert.Nil(t, api)
-				return
-			}
-			assert.NoError(t, err)
-			assert.NotNil(t, api)
-		})
-	}
-}
-
-func TestNewFromAPI_NilReturnsError(t *testing.T) {
-	c, err := NewFromAPI(nil)
-	assert.Nil(t, c)
-	assert.EqualError(t, err, "docker: api client must not be nil")
-}
-
-func TestNewFromAPI_ValidReturnsClient(t *testing.T) {
-	api, err := NewAPIClient("unix:///var/run/docker.sock", nil)
-	assert.NoError(t, err)
-	assert.NotNil(t, api)
-
-	c, err := NewFromAPI(api)
-	assert.NoError(t, err)
-	assert.NotNil(t, c)
-
-	var _ ctr.Client = c
-}
-
-func TestNewClient_DelegatesToHelpers(t *testing.T) {
-	c, err := NewClient("unix:///var/run/docker.sock", nil)
-	assert.NoError(t, err)
-	assert.NotNil(t, c)
-
-	_, err = NewClient("://bad", nil)
-	assert.Error(t, err)
 }
