+++ pkg/runtime/docker/inspect_test.go
@@ -1,6 +1,8 @@
 package docker
 
 import (
+	"context"
+	"errors"
 	"testing"
 
 	ctr "github.com/alexei-led/pumba/pkg/container"
@@ -8,7 +10,106 @@ import (
 	imagetypes "github.com/docker/docker/api/types/image"
 	"github.com/docker/docker/api/types/network"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
 )
+
+func TestListContainers_Success(t *testing.T) {
+	containerDetails := DetailsResponse(AsMap("ID", "foo", "Name", "/bar", "Image", "abc123"))
+	allContainersResponse := Containers(Response(AsMap(
+		"ID", "foo",
+		"Names", []string{"bar"})),
+	)
+	imageDetails := ImageDetailsResponse(AsMap("ID", "abc123"))
+
+	api := NewMockEngine()
+	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(allContainersResponse, nil)
+	api.EXPECT().ContainerInspect(mock.Anything, "foo").Return(containerDetails, nil)
+	api.EXPECT().ImageInspect(mock.Anything, "abc123").Return(imageDetails, nil)
+
+	client := dockerClient{containerAPI: api, imageAPI: api}
+
+	// Verify that dockerClient implements container.Client interface
+	var _ ctr.Client = (*dockerClient)(nil)
+
+	containers, err := client.ListContainers(context.TODO(), mockAllContainers, ctr.ListOpts{All: true})
+
+	assert.NoError(t, err)
+	assert.Len(t, containers, 1)
+	assert.Equal(t, "foo", containers[0].ID())
+	assert.Equal(t, "/bar", containers[0].Name())
+	assert.Equal(t, "abc123", containers[0].ImageID)
+	api.AssertExpectations(t)
+}
+
+func TestListContainers_Filter(t *testing.T) {
+	containerDetails := DetailsResponse(AsMap("ID", "foo", "Name", "/bar", "Image", "abc123"))
+	allContainersResponse := Containers(Response(AsMap(
+		"ID", "foo",
+		"Names", []string{"bar"})),
+	)
+	imageDetails := ImageDetailsResponse(AsMap("ID", "abc123"))
+
+	api := NewMockEngine()
+	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(allContainersResponse, nil)
+	api.EXPECT().ContainerInspect(mock.Anything, "foo").Return(containerDetails, nil)
+	api.EXPECT().ImageInspect(mock.Anything, "abc123").Return(imageDetails, nil)
+
+	client := dockerClient{containerAPI: api, imageAPI: api}
+	containers, err := client.ListContainers(context.TODO(), mockNoContainers, ctr.ListOpts{})
+
+	assert.NoError(t, err)
+	assert.Len(t, containers, 0)
+	api.AssertExpectations(t)
+}
+
+func TestListContainers_ListError(t *testing.T) {
+	api := NewMockEngine()
+	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(Containers(), errors.New("oops"))
+
+	client := dockerClient{containerAPI: api, imageAPI: api}
+	_, err := client.ListContainers(context.TODO(), mockAllContainers, ctr.ListOpts{All: true})
+
+	assert.Error(t, err)
+	assert.EqualError(t, err, "failed to list containers: oops")
+	api.AssertExpectations(t)
+}
+
+func TestListContainers_InspectContainerError(t *testing.T) {
+	api := NewMockEngine()
+	allContainersResponse := Containers(Response(AsMap(
+		"ID", "foo",
+		"Names", []string{"bar"})),
+	)
+	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(allContainersResponse, nil)
+	api.EXPECT().ContainerInspect(mock.Anything, "foo").Return(DetailsResponse(AsMap()), errors.New("uh-oh"))
+
+	client := dockerClient{containerAPI: api, imageAPI: api}
+	_, err := client.ListContainers(context.TODO(), mockAllContainers, ctr.ListOpts{All: true})
+
+	assert.Error(t, err)
+	assert.EqualError(t, err, "failed to inspect container: uh-oh")
+	api.AssertExpectations(t)
+}
+
+func TestListContainers_InspectImageError(t *testing.T) {
+	allContainersResponse := Containers(Response(AsMap(
+		"ID", "foo",
+		"Names", []string{"bar"})),
+	)
+	resp := DetailsResponse(AsMap("Image", "abc123"))
+	imageDetailsResponse := ImageDetailsResponse(AsMap())
+	api := NewMockEngine()
+	api.EXPECT().ContainerList(mock.Anything, mock.Anything).Return(allContainersResponse, nil)
+	api.EXPECT().ContainerInspect(mock.Anything, "foo").Return(resp, nil)
+	api.EXPECT().ImageInspect(mock.Anything, "abc123").Return(imageDetailsResponse, errors.New("whoops"))
+
+	client := dockerClient{containerAPI: api, imageAPI: api}
+	containers, err := client.ListContainers(context.TODO(), mockAllContainers, ctr.ListOpts{All: true})
+
+	assert.NoError(t, err)
+	assert.Len(t, containers, 1)
+	api.AssertExpectations(t)
+}
 
 func TestDockerInspectToContainer(t *testing.T) {
 	tests := []struct {
